+++ sdm-iso-front/src/components/AuthBox/AuthBox.jsx
@@ -1,5 +1,5 @@
 import React, { useRef, useState, useEffect } from 'react';
-import useAuth from '../hooks/useAuth';
+import useAuth from '../Security/hooks/useAuth';
 import axios from '../api/axios';
 import { Link, useNavigate } from 'react-router-dom';
