+++ sdm-iso-front/src/components/Security/hooks/useAuth.jsx
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import AuthContext from "../Context/AuthProvider";
+import AuthContext from "../AuthProvider";
 
 
 const useAuth = () => {
