+++ sdm-iso-front/src/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import reportWebVitals from './reportWebVitals';
-import { AuthProvider } from './components/Context/AuthProvider';
+import { AuthProvider } from './components/Security/AuthProvider';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import './index.css';
 import App from './App';
